requirements.txt
streamlit>=1.32.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
numpy>=1.24.0
facenet-pytorch>=2.5.3
opencv-python-headless>=4.8.0
deepface>=0.0.79
tensorflow-cpu>=2.13.0
